derive target repos from all issues for a contributions-like repository

derive target repos from all issues for a contributions-like repository

get_repos_from_all_issues(owner, repo)

Arguments

owner

character(1) github.com repo owner

repo

character(1) repo name

Value

A vector of URLs supplied by contributors, identifying the URL of their contribution's repo.

Note

The python code was written by instructing perplexity.ai on Nov 18 2025.

Examples

rs = get_repos_from_all_issues("bioconductor", "workshop-contributions")
head(rs)
#> [[1]]
#> [[1]]$url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/117"
#> 
#> [[1]]$repository_url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions"
#> 
#> [[1]]$labels_url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/117/labels{/name}"
#> 
#> [[1]]$comments_url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/117/comments"
#> 
#> [[1]]$events_url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/117/events"
#> 
#> [[1]]$html_url
#> [1] "https://github.com/Bioconductor/workshop-contributions/issues/117"
#> 
#> [[1]]$id
#> [1] -864433747
#> 
#> [[1]]$node_id
#> [1] "I_kwDOJTnwvM7Mecmt"
#> 
#> [[1]]$number
#> [1] 117
#> 
#> [[1]]$title
#> [1] "[EuroBioC2025] Streamlining Omics Data Visualization with `iSEE` and `iSEEfier`"
#> 
#> [[1]]$user
#> [[1]]$user$login
#> [1] "bioc-workshop-dev"
#> 
#> [[1]]$user$id
#> [1] 143761862
#> 
#> [[1]]$user$node_id
#> [1] "U_kgDOCJGhxg"
#> 
#> [[1]]$user$avatar_url
#> [1] "https://avatars.githubusercontent.com/u/143761862?v=4"
#> 
#> [[1]]$user$gravatar_id
#> [1] ""
#> 
#> [[1]]$user$url
#> [1] "https://api.github.com/users/bioc-workshop-dev"
#> 
#> [[1]]$user$html_url
#> [1] "https://github.com/bioc-workshop-dev"
#> 
#> [[1]]$user$followers_url
#> [1] "https://api.github.com/users/bioc-workshop-dev/followers"
#> 
#> [[1]]$user$following_url
#> [1] "https://api.github.com/users/bioc-workshop-dev/following{/other_user}"
#> 
#> [[1]]$user$gists_url
#> [1] "https://api.github.com/users/bioc-workshop-dev/gists{/gist_id}"
#> 
#> [[1]]$user$starred_url
#> [1] "https://api.github.com/users/bioc-workshop-dev/starred{/owner}{/repo}"
#> 
#> [[1]]$user$subscriptions_url
#> [1] "https://api.github.com/users/bioc-workshop-dev/subscriptions"
#> 
#> [[1]]$user$organizations_url
#> [1] "https://api.github.com/users/bioc-workshop-dev/orgs"
#> 
#> [[1]]$user$repos_url
#> [1] "https://api.github.com/users/bioc-workshop-dev/repos"
#> 
#> [[1]]$user$events_url
#> [1] "https://api.github.com/users/bioc-workshop-dev/events{/privacy}"
#> 
#> [[1]]$user$received_events_url
#> [1] "https://api.github.com/users/bioc-workshop-dev/received_events"
#> 
#> [[1]]$user$type
#> [1] "User"
#> 
#> [[1]]$user$user_view_type
#> [1] "public"
#> 
#> [[1]]$user$site_admin
#> [1] FALSE
#> 
#> 
#> [[1]]$labels
#> list()
#> 
#> [[1]]$state
#> [1] "open"
#> 
#> [[1]]$locked
#> [1] FALSE
#> 
#> [[1]]$assignee
#> NULL
#> 
#> [[1]]$assignees
#> list()
#> 
#> [[1]]$milestone
#> NULL
#> 
#> [[1]]$comments
#> [1] 1
#> 
#> [[1]]$created_at
#> [1] "2025-09-18T14:02:32Z"
#> 
#> [[1]]$updated_at
#> [1] "2025-09-18T14:02:33Z"
#> 
#> [[1]]$closed_at
#> NULL
#> 
#> [[1]]$author_association
#> [1] "NONE"
#> 
#> [[1]]$type
#> NULL
#> 
#> [[1]]$active_lock_reason
#> NULL
#> 
#> [[1]]$sub_issues_summary
#> [[1]]$sub_issues_summary$total
#> [1] 0
#> 
#> [[1]]$sub_issues_summary$completed
#> [1] 0
#> 
#> [[1]]$sub_issues_summary$percent_completed
#> [1] 0
#> 
#> 
#> [[1]]$issue_dependencies_summary
#> [[1]]$issue_dependencies_summary$blocked_by
#> [1] 0
#> 
#> [[1]]$issue_dependencies_summary$total_blocked_by
#> [1] 0
#> 
#> [[1]]$issue_dependencies_summary$blocking
#> [1] 0
#> 
#> [[1]]$issue_dependencies_summary$total_blocking
#> [1] 0
#> 
#> 
#> [[1]]$body
#> [1] "\n**DO NOT INCLUDE REQUESTS IN THE FIRST COMMENT.**\n**PLEASE POST THIS TEMPLATE UNCHANGED THEN FOLLOW ITS INSTRUCTIONS IN A NEW\nCOMMENT**\n\n# General Notes\nThis repository serves as a mostly automated pipeline for deploying workshops\nto the [Bioconductor Workshop](https://workshop.bioconductor.org).\n\n# Bot Commands\n\nIssues in this repository are monitored to by a bot. All commands need to be\nwritten on the first line of the comment, in a single line.\n\n## Example Docker Request\n\n```\n/request\n  id='tidybioc2023'\n  title='Tidy Transcriptomics'\n  description='For Single-Cell RNA Sequencing Analyses'\n  section='Smorgasbord 2023'\n  source='https://github.com/tidytranscriptomics-workshops/bioc2023_tidytranscriptomics'\n  docker='ghcr.io/tidytranscriptomics-workshops/bioc2023_tidytranscriptomics:latest'\n```\n\nNote. The example `/request` is on multiple lines for readability.\n\n## Additional Workshop Information\n\nExpected Number of Participants: 5\n\nWorkshop Date(s): September 18, 2025 - September 18, 2025\n\nWorkshop Start Time (approx.): 14:00\n\n"
#> 
#> [[1]]$closed_by
#> NULL
#> 
#> [[1]]$reactions
#> [[1]]$reactions$url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/117/reactions"
#> 
#> [[1]]$reactions$total_count
#> [1] 0
#> 
#> [[1]]$reactions$`+1`
#> [1] 0
#> 
#> [[1]]$reactions$`-1`
#> [1] 0
#> 
#> [[1]]$reactions$laugh
#> [1] 0
#> 
#> [[1]]$reactions$hooray
#> [1] 0
#> 
#> [[1]]$reactions$confused
#> [1] 0
#> 
#> [[1]]$reactions$heart
#> [1] 0
#> 
#> [[1]]$reactions$rocket
#> [1] 0
#> 
#> [[1]]$reactions$eyes
#> [1] 0
#> 
#> 
#> [[1]]$timeline_url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/117/timeline"
#> 
#> [[1]]$performed_via_github_app
#> NULL
#> 
#> [[1]]$state_reason
#> NULL
#> 
#> 
#> [[2]]
#> [[2]]$url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/116"
#> 
#> [[2]]$repository_url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions"
#> 
#> [[2]]$labels_url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/116/labels{/name}"
#> 
#> [[2]]$comments_url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/116/comments"
#> 
#> [[2]]$events_url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/116/events"
#> 
#> [[2]]$html_url
#> [1] "https://github.com/Bioconductor/workshop-contributions/issues/116"
#> 
#> [[2]]$id
#> [1] -864438039
#> 
#> [[2]]$node_id
#> [1] "I_kwDOJTnwvM7Mebjp"
#> 
#> [[2]]$number
#> [1] 116
#> 
#> [[2]]$title
#> [1] "[EuroBioc2025] Package Demo: mia"
#> 
#> [[2]]$user
#> [[2]]$user$login
#> [1] "bioc-workshop-dev"
#> 
#> [[2]]$user$id
#> [1] 143761862
#> 
#> [[2]]$user$node_id
#> [1] "U_kgDOCJGhxg"
#> 
#> [[2]]$user$avatar_url
#> [1] "https://avatars.githubusercontent.com/u/143761862?v=4"
#> 
#> [[2]]$user$gravatar_id
#> [1] ""
#> 
#> [[2]]$user$url
#> [1] "https://api.github.com/users/bioc-workshop-dev"
#> 
#> [[2]]$user$html_url
#> [1] "https://github.com/bioc-workshop-dev"
#> 
#> [[2]]$user$followers_url
#> [1] "https://api.github.com/users/bioc-workshop-dev/followers"
#> 
#> [[2]]$user$following_url
#> [1] "https://api.github.com/users/bioc-workshop-dev/following{/other_user}"
#> 
#> [[2]]$user$gists_url
#> [1] "https://api.github.com/users/bioc-workshop-dev/gists{/gist_id}"
#> 
#> [[2]]$user$starred_url
#> [1] "https://api.github.com/users/bioc-workshop-dev/starred{/owner}{/repo}"
#> 
#> [[2]]$user$subscriptions_url
#> [1] "https://api.github.com/users/bioc-workshop-dev/subscriptions"
#> 
#> [[2]]$user$organizations_url
#> [1] "https://api.github.com/users/bioc-workshop-dev/orgs"
#> 
#> [[2]]$user$repos_url
#> [1] "https://api.github.com/users/bioc-workshop-dev/repos"
#> 
#> [[2]]$user$events_url
#> [1] "https://api.github.com/users/bioc-workshop-dev/events{/privacy}"
#> 
#> [[2]]$user$received_events_url
#> [1] "https://api.github.com/users/bioc-workshop-dev/received_events"
#> 
#> [[2]]$user$type
#> [1] "User"
#> 
#> [[2]]$user$user_view_type
#> [1] "public"
#> 
#> [[2]]$user$site_admin
#> [1] FALSE
#> 
#> 
#> [[2]]$labels
#> list()
#> 
#> [[2]]$state
#> [1] "open"
#> 
#> [[2]]$locked
#> [1] FALSE
#> 
#> [[2]]$assignee
#> NULL
#> 
#> [[2]]$assignees
#> list()
#> 
#> [[2]]$milestone
#> NULL
#> 
#> [[2]]$comments
#> [1] 4
#> 
#> [[2]]$created_at
#> [1] "2025-09-18T14:01:32Z"
#> 
#> [[2]]$updated_at
#> [1] "2025-09-18T14:26:58Z"
#> 
#> [[2]]$closed_at
#> NULL
#> 
#> [[2]]$author_association
#> [1] "NONE"
#> 
#> [[2]]$type
#> NULL
#> 
#> [[2]]$active_lock_reason
#> NULL
#> 
#> [[2]]$sub_issues_summary
#> [[2]]$sub_issues_summary$total
#> [1] 0
#> 
#> [[2]]$sub_issues_summary$completed
#> [1] 0
#> 
#> [[2]]$sub_issues_summary$percent_completed
#> [1] 0
#> 
#> 
#> [[2]]$issue_dependencies_summary
#> [[2]]$issue_dependencies_summary$blocked_by
#> [1] 0
#> 
#> [[2]]$issue_dependencies_summary$total_blocked_by
#> [1] 0
#> 
#> [[2]]$issue_dependencies_summary$blocking
#> [1] 0
#> 
#> [[2]]$issue_dependencies_summary$total_blocking
#> [1] 0
#> 
#> 
#> [[2]]$body
#> [1] "\n**DO NOT INCLUDE REQUESTS IN THE FIRST COMMENT.**\n**PLEASE POST THIS TEMPLATE UNCHANGED THEN FOLLOW ITS INSTRUCTIONS IN A NEW\nCOMMENT**\n\n# General Notes\nThis repository serves as a mostly automated pipeline for deploying workshops\nto the [Bioconductor Workshop](https://workshop.bioconductor.org).\n\n# Bot Commands\n\nIssues in this repository are monitored to by a bot. All commands need to be\nwritten on the first line of the comment, in a single line.\n\n## Example Docker Request\n\n```\n/request\n  id='tidybioc2023'\n  title='Tidy Transcriptomics'\n  description='For Single-Cell RNA Sequencing Analyses'\n  section='Smorgasbord 2023'\n  source='https://github.com/tidytranscriptomics-workshops/bioc2023_tidytranscriptomics'\n  docker='ghcr.io/tidytranscriptomics-workshops/bioc2023_tidytranscriptomics:latest'\n```\n\nNote. The example `/request` is on multiple lines for readability.\n\n## Additional Workshop Information\n\nExpected Number of Participants: 0\n\nWorkshop Date(s): September 18, 2025 - September 18, 2025\n\nWorkshop Start Time (approx.): 13:45\n\n"
#> 
#> [[2]]$closed_by
#> NULL
#> 
#> [[2]]$reactions
#> [[2]]$reactions$url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/116/reactions"
#> 
#> [[2]]$reactions$total_count
#> [1] 0
#> 
#> [[2]]$reactions$`+1`
#> [1] 0
#> 
#> [[2]]$reactions$`-1`
#> [1] 0
#> 
#> [[2]]$reactions$laugh
#> [1] 0
#> 
#> [[2]]$reactions$hooray
#> [1] 0
#> 
#> [[2]]$reactions$confused
#> [1] 0
#> 
#> [[2]]$reactions$heart
#> [1] 0
#> 
#> [[2]]$reactions$rocket
#> [1] 0
#> 
#> [[2]]$reactions$eyes
#> [1] 0
#> 
#> 
#> [[2]]$timeline_url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/116/timeline"
#> 
#> [[2]]$performed_via_github_app
#> NULL
#> 
#> [[2]]$state_reason
#> NULL
#> 
#> 
#> [[3]]
#> [[3]]$url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/115"
#> 
#> [[3]]$repository_url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions"
#> 
#> [[3]]$labels_url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/115/labels{/name}"
#> 
#> [[3]]$comments_url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/115/comments"
#> 
#> [[3]]$events_url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/115/events"
#> 
#> [[3]]$html_url
#> [1] "https://github.com/Bioconductor/workshop-contributions/issues/115"
#> 
#> [[3]]$id
#> [1] -875372824
#> 
#> [[3]]$node_id
#> [1] "I_kwDOJTnwvM7L0t7o"
#> 
#> [[3]]$number
#> [1] 115
#> 
#> [[3]]$title
#> [1] "[EuroBioc2025] OSTA workshop"
#> 
#> [[3]]$user
#> [[3]]$user$login
#> [1] "almahmoud"
#> 
#> [[3]]$user$id
#> [1] 14061824
#> 
#> [[3]]$user$node_id
#> [1] "MDQ6VXNlcjE0MDYxODI0"
#> 
#> [[3]]$user$avatar_url
#> [1] "https://avatars.githubusercontent.com/u/14061824?v=4"
#> 
#> [[3]]$user$gravatar_id
#> [1] ""
#> 
#> [[3]]$user$url
#> [1] "https://api.github.com/users/almahmoud"
#> 
#> [[3]]$user$html_url
#> [1] "https://github.com/almahmoud"
#> 
#> [[3]]$user$followers_url
#> [1] "https://api.github.com/users/almahmoud/followers"
#> 
#> [[3]]$user$following_url
#> [1] "https://api.github.com/users/almahmoud/following{/other_user}"
#> 
#> [[3]]$user$gists_url
#> [1] "https://api.github.com/users/almahmoud/gists{/gist_id}"
#> 
#> [[3]]$user$starred_url
#> [1] "https://api.github.com/users/almahmoud/starred{/owner}{/repo}"
#> 
#> [[3]]$user$subscriptions_url
#> [1] "https://api.github.com/users/almahmoud/subscriptions"
#> 
#> [[3]]$user$organizations_url
#> [1] "https://api.github.com/users/almahmoud/orgs"
#> 
#> [[3]]$user$repos_url
#> [1] "https://api.github.com/users/almahmoud/repos"
#> 
#> [[3]]$user$events_url
#> [1] "https://api.github.com/users/almahmoud/events{/privacy}"
#> 
#> [[3]]$user$received_events_url
#> [1] "https://api.github.com/users/almahmoud/received_events"
#> 
#> [[3]]$user$type
#> [1] "User"
#> 
#> [[3]]$user$user_view_type
#> [1] "public"
#> 
#> [[3]]$user$site_admin
#> [1] FALSE
#> 
#> 
#> [[3]]$labels
#> [[3]]$labels[[1]]
#> [[3]]$labels[[1]]$id
#> [1] -1372889707
#> 
#> [[3]]$labels[[1]]$node_id
#> [1] "LA_kwDOJTnwvM8AAAABritdlQ"
#> 
#> [[3]]$labels[[1]]$url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/labels/deployed"
#> 
#> [[3]]$labels[[1]]$name
#> [1] "deployed"
#> 
#> [[3]]$labels[[1]]$color
#> [1] "9541A4"
#> 
#> [[3]]$labels[[1]]$default
#> [1] FALSE
#> 
#> [[3]]$labels[[1]]$description
#> [1] ""
#> 
#> 
#> 
#> [[3]]$state
#> [1] "open"
#> 
#> [[3]]$locked
#> [1] FALSE
#> 
#> [[3]]$assignee
#> NULL
#> 
#> [[3]]$assignees
#> list()
#> 
#> [[3]]$milestone
#> NULL
#> 
#> [[3]]$comments
#> [1] 6
#> 
#> [[3]]$created_at
#> [1] "2025-09-15T21:32:05Z"
#> 
#> [[3]]$updated_at
#> [1] "2025-09-18T08:13:59Z"
#> 
#> [[3]]$closed_at
#> NULL
#> 
#> [[3]]$author_association
#> [1] "MEMBER"
#> 
#> [[3]]$type
#> NULL
#> 
#> [[3]]$active_lock_reason
#> NULL
#> 
#> [[3]]$sub_issues_summary
#> [[3]]$sub_issues_summary$total
#> [1] 0
#> 
#> [[3]]$sub_issues_summary$completed
#> [1] 0
#> 
#> [[3]]$sub_issues_summary$percent_completed
#> [1] 0
#> 
#> 
#> [[3]]$issue_dependencies_summary
#> [[3]]$issue_dependencies_summary$blocked_by
#> [1] 0
#> 
#> [[3]]$issue_dependencies_summary$total_blocked_by
#> [1] 0
#> 
#> [[3]]$issue_dependencies_summary$blocking
#> [1] 0
#> 
#> [[3]]$issue_dependencies_summary$total_blocking
#> [1] 0
#> 
#> 
#> [[3]]$body
#> [1] "**DO NOT INCLUDE REQUESTS IN THE FIRST COMMENT.**\n**PLEASE POST THIS TEMPLATE UNCHANGED THEN FOLLOW ITS INSTRUCTIONS IN A NEW COMMENT**\n\n\n# General Notes\nThis repository serves as a mostly automated pipeline for deploying workshops to the [Bioconductor Workshop](https://workshop.bioconductor.org).\n\n# Bot Commands\n\nIssues in this repository are listen to by a bot. All commands need to be written on the first line of the comment, in a single line, starting with the `/command` and containing all `parameter=\"value in here\"` pairs.\n\n## Public request\nUnless you are a member of the Bioconductor Core team or an administrator of this repository, you can only perform `/request` commands in this space. Some parameters are universal to all requests and mandatory, namely the Title, Description, Source URL, and Section on the workshop instance. You may request your workshop be added to an existing section, or request a new Section name.\nBelow is an example of an incomplete request with all mandatory parameters:\n```\n/request id=\"myworkshopuniqueid\" title=\"Bold Text\" description=\"unbolded text next to the title\" section=\"Conference 1996\" source=\"https://github.com/super/repository\" \n```\n\nThis request will then need to be completed with an additional 1-2 parameters depending on your source for the workshop.\n\nWe currently support 3 types of submissions:\n\n1) `docker=` parameter for pre-built RStudio containers (such as Orchestra workshops built based on [BuildABiocWorkshop](https://github.com/Bioconductor/BuildABiocWorkshop))\nBelow is an example of a full `/request` command for a workshop with a pre-built docker container:\n```\n/request id=\"tidybioc2022\" title=\"Tidy Transcriptomics\" description=\"For Single-Cell RNA Sequencing Analyses\" section=\"Smorgasbord 2023\" source=\"https://github.com/tidytranscriptomics-workshops/bioc2022_tidytranscriptomics\" docker=\"ghcr.io/tidytranscriptomics-workshops/bioc2022_tidytranscriptomics:latest\"\n```\n\n2) `vignettes=` parameter for Rmd/qmd workshops.\n\nNote that when any `Rmd` or `qmd` files are pulled as part of a request including vignettes, the script will only deploy successfully if the vignettes can successfully render. In the process, all dependencies will be automatically scraped and installed in the resulting container image, making it optional to explicitly mention dependencies.\n\n2) a) In-source vignettes, especially useful for example if your Rmd references other files (eg: images), in which case you should include them in your source repository. Parameter represents a comma-separated list of relative paths (with wildcards accepted) from your source repository. Eg: `vignettes=\"vignettes/*,images/*\"` will copy all files from your source repository under those two subdirectories.\n\nBelow is an example of a full request with in-source vignettes:\n```\n/request id=\"annotation316\" title=\"Genomic Annotation Resources\" description=\"with Bioconductor annotation package\" section=\"Smorgasbord 2023\" source=\"https://github.com/Bioconductor/annotation\" vignettes=\"vignettes/*\"\n```\n\n2) b) markdown file urls, in the form of a comma-separated list of URLs. eg: `vignettes=\"https://raw.githubusercontent.com/Bioconductor/annotation/devel/vignettes/Annotation_Resources.Rmd,https://raw.githubusercontent.com/Bioconductor/annotation/devel/vignettes/Annotating_Genomic_Ranges.Rmd\"`. This can especially be useful for people unfamiliar with github, who could write a self-contained Rmd file in RStudio, paste it at https://gist.github.com where single files can be hosted with no directory structure or `git` operations.\n\nBelow is an example of a full request for a url vignette:\n```\n/request id=\"genranges316\" title=\"Annotating Genomic Ranges\" description=\"with Bioconductor annotation package\" section=\"Smorgasbord 2023\" source=\"https://raw.githubusercontent.com/Bioconductor/annotation/devel/vignettes/Annotating_Genomic_Ranges.Rmd\" vignettes=\"https://raw.githubusercontent.com/Bioconductor/annotation/devel/vignettes/Annotating_Genomic_Ranges.Rmd,https://raw.githubusercontent.com/Bioconductor/annotation/devel/vignettes/Annotating_Genomic_Ranges.Rmd\"\n```\n\n3) `pkglist=` which can be used with or without the `vignettes=` option, in order to add a list of Bioconductor and/or CRAN packages to the resulting auto-built container image for this request. It should be passed as a comma-separated list of package names. eg: `pkglist=\"VariantAnnotation,AnnotationHub,TxDb.Hsapiens.UCSC.hg19.knownGene\"`\n\nBelow is an example of a request to add a package demo\n\n```\n/request id=\"bsgenome316\" title=\"BSgenome package demo\" description=\"in the latest Bioconductor\" section=\"Package Demos\" source=\"https://github.com/Bioconductor/BSgenome\" pkglist=\"BSgenome\"\n```\n\nEvery request must contain at least 1 of the 3 listed parameters, in addition to the mandatory parameters listed above.\n\nYou may add any comments, information or special requests after the second line, but the first line must contain only the `/command` and its parameters.\n\n## Admin request\nBioconductor Core Team and repository admins will be able to deploy vetted requests to one of two servers. These requests take the same parameters as the `/request` command, but go through with building and deploying the changes. When the request does not include a `docker=` parameter, the image building might delay a response by a couple of hours for package-intensive workshops.\n\nThe `/test` command will deploy the instance to our test server, where the requester can verify the aesthetics as well as functionality of the workshop and request any changes.\n\nThe `/publish` command will deploy the final approved workshop to the production instance.\n\nThe best way for an admin to populate these requests is copy-paste the last passing full `/request` command and all parameters from the requester, and replace the `/request` command with the appropriate deployment directive.\n"
#> 
#> [[3]]$closed_by
#> NULL
#> 
#> [[3]]$reactions
#> [[3]]$reactions$url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/115/reactions"
#> 
#> [[3]]$reactions$total_count
#> [1] 0
#> 
#> [[3]]$reactions$`+1`
#> [1] 0
#> 
#> [[3]]$reactions$`-1`
#> [1] 0
#> 
#> [[3]]$reactions$laugh
#> [1] 0
#> 
#> [[3]]$reactions$hooray
#> [1] 0
#> 
#> [[3]]$reactions$confused
#> [1] 0
#> 
#> [[3]]$reactions$heart
#> [1] 0
#> 
#> [[3]]$reactions$rocket
#> [1] 0
#> 
#> [[3]]$reactions$eyes
#> [1] 0
#> 
#> 
#> [[3]]$timeline_url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/115/timeline"
#> 
#> [[3]]$performed_via_github_app
#> NULL
#> 
#> [[3]]$state_reason
#> NULL
#> 
#> 
#> [[4]]
#> [[4]]$url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/114"
#> 
#> [[4]]$repository_url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions"
#> 
#> [[4]]$labels_url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/114/labels{/name}"
#> 
#> [[4]]$comments_url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/114/comments"
#> 
#> [[4]]$events_url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/114/events"
#> 
#> [[4]]$html_url
#> [1] "https://github.com/Bioconductor/workshop-contributions/issues/114"
#> 
#> [[4]]$id
#> [1] -875383011
#> 
#> [[4]]$node_id
#> [1] "I_kwDOJTnwvM7L0rcd"
#> 
#> [[4]]$number
#> [1] 114
#> 
#> [[4]]$title
#> [1] "[EuroBioc2025] iSEEWorkshop"
#> 
#> [[4]]$user
#> [[4]]$user$login
#> [1] "almahmoud"
#> 
#> [[4]]$user$id
#> [1] 14061824
#> 
#> [[4]]$user$node_id
#> [1] "MDQ6VXNlcjE0MDYxODI0"
#> 
#> [[4]]$user$avatar_url
#> [1] "https://avatars.githubusercontent.com/u/14061824?v=4"
#> 
#> [[4]]$user$gravatar_id
#> [1] ""
#> 
#> [[4]]$user$url
#> [1] "https://api.github.com/users/almahmoud"
#> 
#> [[4]]$user$html_url
#> [1] "https://github.com/almahmoud"
#> 
#> [[4]]$user$followers_url
#> [1] "https://api.github.com/users/almahmoud/followers"
#> 
#> [[4]]$user$following_url
#> [1] "https://api.github.com/users/almahmoud/following{/other_user}"
#> 
#> [[4]]$user$gists_url
#> [1] "https://api.github.com/users/almahmoud/gists{/gist_id}"
#> 
#> [[4]]$user$starred_url
#> [1] "https://api.github.com/users/almahmoud/starred{/owner}{/repo}"
#> 
#> [[4]]$user$subscriptions_url
#> [1] "https://api.github.com/users/almahmoud/subscriptions"
#> 
#> [[4]]$user$organizations_url
#> [1] "https://api.github.com/users/almahmoud/orgs"
#> 
#> [[4]]$user$repos_url
#> [1] "https://api.github.com/users/almahmoud/repos"
#> 
#> [[4]]$user$events_url
#> [1] "https://api.github.com/users/almahmoud/events{/privacy}"
#> 
#> [[4]]$user$received_events_url
#> [1] "https://api.github.com/users/almahmoud/received_events"
#> 
#> [[4]]$user$type
#> [1] "User"
#> 
#> [[4]]$user$user_view_type
#> [1] "public"
#> 
#> [[4]]$user$site_admin
#> [1] FALSE
#> 
#> 
#> [[4]]$labels
#> [[4]]$labels[[1]]
#> [[4]]$labels[[1]]$id
#> [1] -1372889707
#> 
#> [[4]]$labels[[1]]$node_id
#> [1] "LA_kwDOJTnwvM8AAAABritdlQ"
#> 
#> [[4]]$labels[[1]]$url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/labels/deployed"
#> 
#> [[4]]$labels[[1]]$name
#> [1] "deployed"
#> 
#> [[4]]$labels[[1]]$color
#> [1] "9541A4"
#> 
#> [[4]]$labels[[1]]$default
#> [1] FALSE
#> 
#> [[4]]$labels[[1]]$description
#> [1] ""
#> 
#> 
#> 
#> [[4]]$state
#> [1] "open"
#> 
#> [[4]]$locked
#> [1] FALSE
#> 
#> [[4]]$assignee
#> NULL
#> 
#> [[4]]$assignees
#> list()
#> 
#> [[4]]$milestone
#> NULL
#> 
#> [[4]]$comments
#> [1] 3
#> 
#> [[4]]$created_at
#> [1] "2025-09-15T21:28:00Z"
#> 
#> [[4]]$updated_at
#> [1] "2025-09-17T08:05:37Z"
#> 
#> [[4]]$closed_at
#> NULL
#> 
#> [[4]]$author_association
#> [1] "MEMBER"
#> 
#> [[4]]$type
#> NULL
#> 
#> [[4]]$active_lock_reason
#> NULL
#> 
#> [[4]]$sub_issues_summary
#> [[4]]$sub_issues_summary$total
#> [1] 0
#> 
#> [[4]]$sub_issues_summary$completed
#> [1] 0
#> 
#> [[4]]$sub_issues_summary$percent_completed
#> [1] 0
#> 
#> 
#> [[4]]$issue_dependencies_summary
#> [[4]]$issue_dependencies_summary$blocked_by
#> [1] 0
#> 
#> [[4]]$issue_dependencies_summary$total_blocked_by
#> [1] 0
#> 
#> [[4]]$issue_dependencies_summary$blocking
#> [1] 0
#> 
#> [[4]]$issue_dependencies_summary$total_blocking
#> [1] 0
#> 
#> 
#> [[4]]$body
#> [1] "**DO NOT INCLUDE REQUESTS IN THE FIRST COMMENT.**\n**PLEASE POST THIS TEMPLATE UNCHANGED THEN FOLLOW ITS INSTRUCTIONS IN A NEW COMMENT**\n\n\n# General Notes\nThis repository serves as a mostly automated pipeline for deploying workshops to the [Bioconductor Workshop](https://workshop.bioconductor.org).\n\n# Bot Commands\n\nIssues in this repository are listen to by a bot. All commands need to be written on the first line of the comment, in a single line, starting with the `/command` and containing all `parameter=\"value in here\"` pairs.\n\n## Public request\nUnless you are a member of the Bioconductor Core team or an administrator of this repository, you can only perform `/request` commands in this space. Some parameters are universal to all requests and mandatory, namely the Title, Description, Source URL, and Section on the workshop instance. You may request your workshop be added to an existing section, or request a new Section name.\nBelow is an example of an incomplete request with all mandatory parameters:\n```\n/request id=\"myworkshopuniqueid\" title=\"Bold Text\" description=\"unbolded text next to the title\" section=\"Conference 1996\" source=\"https://github.com/super/repository\" \n```\n\nThis request will then need to be completed with an additional 1-2 parameters depending on your source for the workshop.\n\nWe currently support 3 types of submissions:\n\n1) `docker=` parameter for pre-built RStudio containers (such as Orchestra workshops built based on [BuildABiocWorkshop](https://github.com/Bioconductor/BuildABiocWorkshop))\nBelow is an example of a full `/request` command for a workshop with a pre-built docker container:\n```\n/request id=\"tidybioc2022\" title=\"Tidy Transcriptomics\" description=\"For Single-Cell RNA Sequencing Analyses\" section=\"Smorgasbord 2023\" source=\"https://github.com/tidytranscriptomics-workshops/bioc2022_tidytranscriptomics\" docker=\"ghcr.io/tidytranscriptomics-workshops/bioc2022_tidytranscriptomics:latest\"\n```\n\n2) `vignettes=` parameter for Rmd/qmd workshops.\n\nNote that when any `Rmd` or `qmd` files are pulled as part of a request including vignettes, the script will only deploy successfully if the vignettes can successfully render. In the process, all dependencies will be automatically scraped and installed in the resulting container image, making it optional to explicitly mention dependencies.\n\n2) a) In-source vignettes, especially useful for example if your Rmd references other files (eg: images), in which case you should include them in your source repository. Parameter represents a comma-separated list of relative paths (with wildcards accepted) from your source repository. Eg: `vignettes=\"vignettes/*,images/*\"` will copy all files from your source repository under those two subdirectories.\n\nBelow is an example of a full request with in-source vignettes:\n```\n/request id=\"annotation316\" title=\"Genomic Annotation Resources\" description=\"with Bioconductor annotation package\" section=\"Smorgasbord 2023\" source=\"https://github.com/Bioconductor/annotation\" vignettes=\"vignettes/*\"\n```\n\n2) b) markdown file urls, in the form of a comma-separated list of URLs. eg: `vignettes=\"https://raw.githubusercontent.com/Bioconductor/annotation/devel/vignettes/Annotation_Resources.Rmd,https://raw.githubusercontent.com/Bioconductor/annotation/devel/vignettes/Annotating_Genomic_Ranges.Rmd\"`. This can especially be useful for people unfamiliar with github, who could write a self-contained Rmd file in RStudio, paste it at https://gist.github.com where single files can be hosted with no directory structure or `git` operations.\n\nBelow is an example of a full request for a url vignette:\n```\n/request id=\"genranges316\" title=\"Annotating Genomic Ranges\" description=\"with Bioconductor annotation package\" section=\"Smorgasbord 2023\" source=\"https://raw.githubusercontent.com/Bioconductor/annotation/devel/vignettes/Annotating_Genomic_Ranges.Rmd\" vignettes=\"https://raw.githubusercontent.com/Bioconductor/annotation/devel/vignettes/Annotating_Genomic_Ranges.Rmd,https://raw.githubusercontent.com/Bioconductor/annotation/devel/vignettes/Annotating_Genomic_Ranges.Rmd\"\n```\n\n3) `pkglist=` which can be used with or without the `vignettes=` option, in order to add a list of Bioconductor and/or CRAN packages to the resulting auto-built container image for this request. It should be passed as a comma-separated list of package names. eg: `pkglist=\"VariantAnnotation,AnnotationHub,TxDb.Hsapiens.UCSC.hg19.knownGene\"`\n\nBelow is an example of a request to add a package demo\n\n```\n/request id=\"bsgenome316\" title=\"BSgenome package demo\" description=\"in the latest Bioconductor\" section=\"Package Demos\" source=\"https://github.com/Bioconductor/BSgenome\" pkglist=\"BSgenome\"\n```\n\nEvery request must contain at least 1 of the 3 listed parameters, in addition to the mandatory parameters listed above.\n\nYou may add any comments, information or special requests after the second line, but the first line must contain only the `/command` and its parameters.\n\n## Admin request\nBioconductor Core Team and repository admins will be able to deploy vetted requests to one of two servers. These requests take the same parameters as the `/request` command, but go through with building and deploying the changes. When the request does not include a `docker=` parameter, the image building might delay a response by a couple of hours for package-intensive workshops.\n\nThe `/test` command will deploy the instance to our test server, where the requester can verify the aesthetics as well as functionality of the workshop and request any changes.\n\nThe `/publish` command will deploy the final approved workshop to the production instance.\n\nThe best way for an admin to populate these requests is copy-paste the last passing full `/request` command and all parameters from the requester, and replace the `/request` command with the appropriate deployment directive.\n"
#> 
#> [[4]]$closed_by
#> NULL
#> 
#> [[4]]$reactions
#> [[4]]$reactions$url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/114/reactions"
#> 
#> [[4]]$reactions$total_count
#> [1] 0
#> 
#> [[4]]$reactions$`+1`
#> [1] 0
#> 
#> [[4]]$reactions$`-1`
#> [1] 0
#> 
#> [[4]]$reactions$laugh
#> [1] 0
#> 
#> [[4]]$reactions$hooray
#> [1] 0
#> 
#> [[4]]$reactions$confused
#> [1] 0
#> 
#> [[4]]$reactions$heart
#> [1] 0
#> 
#> [[4]]$reactions$rocket
#> [1] 0
#> 
#> [[4]]$reactions$eyes
#> [1] 0
#> 
#> 
#> [[4]]$timeline_url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/114/timeline"
#> 
#> [[4]]$performed_via_github_app
#> NULL
#> 
#> [[4]]$state_reason
#> NULL
#> 
#> 
#> [[5]]
#> [[5]]$url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/113"
#> 
#> [[5]]$repository_url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions"
#> 
#> [[5]]$labels_url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/113/labels{/name}"
#> 
#> [[5]]$comments_url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/113/comments"
#> 
#> [[5]]$events_url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/113/events"
#> 
#> [[5]]$html_url
#> [1] "https://github.com/Bioconductor/workshop-contributions/issues/113"
#> 
#> [[5]]$id
#> [1] -875395291
#> 
#> [[5]]$node_id
#> [1] "I_kwDOJTnwvM7L0ocl"
#> 
#> [[5]]$number
#> [1] 113
#> 
#> [[5]]$title
#> [1] "[EuroBioc2025] msqrob2"
#> 
#> [[5]]$user
#> [[5]]$user$login
#> [1] "almahmoud"
#> 
#> [[5]]$user$id
#> [1] 14061824
#> 
#> [[5]]$user$node_id
#> [1] "MDQ6VXNlcjE0MDYxODI0"
#> 
#> [[5]]$user$avatar_url
#> [1] "https://avatars.githubusercontent.com/u/14061824?v=4"
#> 
#> [[5]]$user$gravatar_id
#> [1] ""
#> 
#> [[5]]$user$url
#> [1] "https://api.github.com/users/almahmoud"
#> 
#> [[5]]$user$html_url
#> [1] "https://github.com/almahmoud"
#> 
#> [[5]]$user$followers_url
#> [1] "https://api.github.com/users/almahmoud/followers"
#> 
#> [[5]]$user$following_url
#> [1] "https://api.github.com/users/almahmoud/following{/other_user}"
#> 
#> [[5]]$user$gists_url
#> [1] "https://api.github.com/users/almahmoud/gists{/gist_id}"
#> 
#> [[5]]$user$starred_url
#> [1] "https://api.github.com/users/almahmoud/starred{/owner}{/repo}"
#> 
#> [[5]]$user$subscriptions_url
#> [1] "https://api.github.com/users/almahmoud/subscriptions"
#> 
#> [[5]]$user$organizations_url
#> [1] "https://api.github.com/users/almahmoud/orgs"
#> 
#> [[5]]$user$repos_url
#> [1] "https://api.github.com/users/almahmoud/repos"
#> 
#> [[5]]$user$events_url
#> [1] "https://api.github.com/users/almahmoud/events{/privacy}"
#> 
#> [[5]]$user$received_events_url
#> [1] "https://api.github.com/users/almahmoud/received_events"
#> 
#> [[5]]$user$type
#> [1] "User"
#> 
#> [[5]]$user$user_view_type
#> [1] "public"
#> 
#> [[5]]$user$site_admin
#> [1] FALSE
#> 
#> 
#> [[5]]$labels
#> [[5]]$labels[[1]]
#> [[5]]$labels[[1]]$id
#> [1] -1372889707
#> 
#> [[5]]$labels[[1]]$node_id
#> [1] "LA_kwDOJTnwvM8AAAABritdlQ"
#> 
#> [[5]]$labels[[1]]$url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/labels/deployed"
#> 
#> [[5]]$labels[[1]]$name
#> [1] "deployed"
#> 
#> [[5]]$labels[[1]]$color
#> [1] "9541A4"
#> 
#> [[5]]$labels[[1]]$default
#> [1] FALSE
#> 
#> [[5]]$labels[[1]]$description
#> [1] ""
#> 
#> 
#> 
#> [[5]]$state
#> [1] "open"
#> 
#> [[5]]$locked
#> [1] FALSE
#> 
#> [[5]]$assignee
#> NULL
#> 
#> [[5]]$assignees
#> list()
#> 
#> [[5]]$milestone
#> NULL
#> 
#> [[5]]$comments
#> [1] 3
#> 
#> [[5]]$created_at
#> [1] "2025-09-15T21:22:57Z"
#> 
#> [[5]]$updated_at
#> [1] "2025-09-17T08:05:33Z"
#> 
#> [[5]]$closed_at
#> NULL
#> 
#> [[5]]$author_association
#> [1] "MEMBER"
#> 
#> [[5]]$type
#> NULL
#> 
#> [[5]]$active_lock_reason
#> NULL
#> 
#> [[5]]$sub_issues_summary
#> [[5]]$sub_issues_summary$total
#> [1] 0
#> 
#> [[5]]$sub_issues_summary$completed
#> [1] 0
#> 
#> [[5]]$sub_issues_summary$percent_completed
#> [1] 0
#> 
#> 
#> [[5]]$issue_dependencies_summary
#> [[5]]$issue_dependencies_summary$blocked_by
#> [1] 0
#> 
#> [[5]]$issue_dependencies_summary$total_blocked_by
#> [1] 0
#> 
#> [[5]]$issue_dependencies_summary$blocking
#> [1] 0
#> 
#> [[5]]$issue_dependencies_summary$total_blocking
#> [1] 0
#> 
#> 
#> [[5]]$body
#> [1] "**DO NOT INCLUDE REQUESTS IN THE FIRST COMMENT.**\n**PLEASE POST THIS TEMPLATE UNCHANGED THEN FOLLOW ITS INSTRUCTIONS IN A NEW COMMENT**\n\n\n# General Notes\nThis repository serves as a mostly automated pipeline for deploying workshops to the [Bioconductor Workshop](https://workshop.bioconductor.org).\n\n# Bot Commands\n\nIssues in this repository are listen to by a bot. All commands need to be written on the first line of the comment, in a single line, starting with the `/command` and containing all `parameter=\"value in here\"` pairs.\n\n## Public request\nUnless you are a member of the Bioconductor Core team or an administrator of this repository, you can only perform `/request` commands in this space. Some parameters are universal to all requests and mandatory, namely the Title, Description, Source URL, and Section on the workshop instance. You may request your workshop be added to an existing section, or request a new Section name.\nBelow is an example of an incomplete request with all mandatory parameters:\n```\n/request id=\"myworkshopuniqueid\" title=\"Bold Text\" description=\"unbolded text next to the title\" section=\"Conference 1996\" source=\"https://github.com/super/repository\" \n```\n\nThis request will then need to be completed with an additional 1-2 parameters depending on your source for the workshop.\n\nWe currently support 3 types of submissions:\n\n1) `docker=` parameter for pre-built RStudio containers (such as Orchestra workshops built based on [BuildABiocWorkshop](https://github.com/Bioconductor/BuildABiocWorkshop))\nBelow is an example of a full `/request` command for a workshop with a pre-built docker container:\n```\n/request id=\"tidybioc2022\" title=\"Tidy Transcriptomics\" description=\"For Single-Cell RNA Sequencing Analyses\" section=\"Smorgasbord 2023\" source=\"https://github.com/tidytranscriptomics-workshops/bioc2022_tidytranscriptomics\" docker=\"ghcr.io/tidytranscriptomics-workshops/bioc2022_tidytranscriptomics:latest\"\n```\n\n2) `vignettes=` parameter for Rmd/qmd workshops.\n\nNote that when any `Rmd` or `qmd` files are pulled as part of a request including vignettes, the script will only deploy successfully if the vignettes can successfully render. In the process, all dependencies will be automatically scraped and installed in the resulting container image, making it optional to explicitly mention dependencies.\n\n2) a) In-source vignettes, especially useful for example if your Rmd references other files (eg: images), in which case you should include them in your source repository. Parameter represents a comma-separated list of relative paths (with wildcards accepted) from your source repository. Eg: `vignettes=\"vignettes/*,images/*\"` will copy all files from your source repository under those two subdirectories.\n\nBelow is an example of a full request with in-source vignettes:\n```\n/request id=\"annotation316\" title=\"Genomic Annotation Resources\" description=\"with Bioconductor annotation package\" section=\"Smorgasbord 2023\" source=\"https://github.com/Bioconductor/annotation\" vignettes=\"vignettes/*\"\n```\n\n2) b) markdown file urls, in the form of a comma-separated list of URLs. eg: `vignettes=\"https://raw.githubusercontent.com/Bioconductor/annotation/devel/vignettes/Annotation_Resources.Rmd,https://raw.githubusercontent.com/Bioconductor/annotation/devel/vignettes/Annotating_Genomic_Ranges.Rmd\"`. This can especially be useful for people unfamiliar with github, who could write a self-contained Rmd file in RStudio, paste it at https://gist.github.com where single files can be hosted with no directory structure or `git` operations.\n\nBelow is an example of a full request for a url vignette:\n```\n/request id=\"genranges316\" title=\"Annotating Genomic Ranges\" description=\"with Bioconductor annotation package\" section=\"Smorgasbord 2023\" source=\"https://raw.githubusercontent.com/Bioconductor/annotation/devel/vignettes/Annotating_Genomic_Ranges.Rmd\" vignettes=\"https://raw.githubusercontent.com/Bioconductor/annotation/devel/vignettes/Annotating_Genomic_Ranges.Rmd,https://raw.githubusercontent.com/Bioconductor/annotation/devel/vignettes/Annotating_Genomic_Ranges.Rmd\"\n```\n\n3) `pkglist=` which can be used with or without the `vignettes=` option, in order to add a list of Bioconductor and/or CRAN packages to the resulting auto-built container image for this request. It should be passed as a comma-separated list of package names. eg: `pkglist=\"VariantAnnotation,AnnotationHub,TxDb.Hsapiens.UCSC.hg19.knownGene\"`\n\nBelow is an example of a request to add a package demo\n\n```\n/request id=\"bsgenome316\" title=\"BSgenome package demo\" description=\"in the latest Bioconductor\" section=\"Package Demos\" source=\"https://github.com/Bioconductor/BSgenome\" pkglist=\"BSgenome\"\n```\n\nEvery request must contain at least 1 of the 3 listed parameters, in addition to the mandatory parameters listed above.\n\nYou may add any comments, information or special requests after the second line, but the first line must contain only the `/command` and its parameters.\n\n## Admin request\nBioconductor Core Team and repository admins will be able to deploy vetted requests to one of two servers. These requests take the same parameters as the `/request` command, but go through with building and deploying the changes. When the request does not include a `docker=` parameter, the image building might delay a response by a couple of hours for package-intensive workshops.\n\nThe `/test` command will deploy the instance to our test server, where the requester can verify the aesthetics as well as functionality of the workshop and request any changes.\n\nThe `/publish` command will deploy the final approved workshop to the production instance.\n\nThe best way for an admin to populate these requests is copy-paste the last passing full `/request` command and all parameters from the requester, and replace the `/request` command with the appropriate deployment directive.\n"
#> 
#> [[5]]$closed_by
#> NULL
#> 
#> [[5]]$reactions
#> [[5]]$reactions$url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/113/reactions"
#> 
#> [[5]]$reactions$total_count
#> [1] 0
#> 
#> [[5]]$reactions$`+1`
#> [1] 0
#> 
#> [[5]]$reactions$`-1`
#> [1] 0
#> 
#> [[5]]$reactions$laugh
#> [1] 0
#> 
#> [[5]]$reactions$hooray
#> [1] 0
#> 
#> [[5]]$reactions$confused
#> [1] 0
#> 
#> [[5]]$reactions$heart
#> [1] 0
#> 
#> [[5]]$reactions$rocket
#> [1] 0
#> 
#> [[5]]$reactions$eyes
#> [1] 0
#> 
#> 
#> [[5]]$timeline_url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/113/timeline"
#> 
#> [[5]]$performed_via_github_app
#> NULL
#> 
#> [[5]]$state_reason
#> NULL
#> 
#> 
#> [[6]]
#> [[6]]$url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/112"
#> 
#> [[6]]$repository_url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions"
#> 
#> [[6]]$labels_url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/112/labels{/name}"
#> 
#> [[6]]$comments_url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/112/comments"
#> 
#> [[6]]$events_url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/112/events"
#> 
#> [[6]]$html_url
#> [1] "https://github.com/Bioconductor/workshop-contributions/issues/112"
#> 
#> [[6]]$id
#> [1] -876122936
#> 
#> [[6]]$node_id
#> [1] "I_kwDOJTnwvM7Lx2zI"
#> 
#> [[6]]$number
#> [1] 112
#> 
#> [[6]]$title
#> [1] "[EuroBioc2025] Orchestrating Microbiome Analysis with Bioconductor"
#> 
#> [[6]]$user
#> [[6]]$user$login
#> [1] "almahmoud"
#> 
#> [[6]]$user$id
#> [1] 14061824
#> 
#> [[6]]$user$node_id
#> [1] "MDQ6VXNlcjE0MDYxODI0"
#> 
#> [[6]]$user$avatar_url
#> [1] "https://avatars.githubusercontent.com/u/14061824?v=4"
#> 
#> [[6]]$user$gravatar_id
#> [1] ""
#> 
#> [[6]]$user$url
#> [1] "https://api.github.com/users/almahmoud"
#> 
#> [[6]]$user$html_url
#> [1] "https://github.com/almahmoud"
#> 
#> [[6]]$user$followers_url
#> [1] "https://api.github.com/users/almahmoud/followers"
#> 
#> [[6]]$user$following_url
#> [1] "https://api.github.com/users/almahmoud/following{/other_user}"
#> 
#> [[6]]$user$gists_url
#> [1] "https://api.github.com/users/almahmoud/gists{/gist_id}"
#> 
#> [[6]]$user$starred_url
#> [1] "https://api.github.com/users/almahmoud/starred{/owner}{/repo}"
#> 
#> [[6]]$user$subscriptions_url
#> [1] "https://api.github.com/users/almahmoud/subscriptions"
#> 
#> [[6]]$user$organizations_url
#> [1] "https://api.github.com/users/almahmoud/orgs"
#> 
#> [[6]]$user$repos_url
#> [1] "https://api.github.com/users/almahmoud/repos"
#> 
#> [[6]]$user$events_url
#> [1] "https://api.github.com/users/almahmoud/events{/privacy}"
#> 
#> [[6]]$user$received_events_url
#> [1] "https://api.github.com/users/almahmoud/received_events"
#> 
#> [[6]]$user$type
#> [1] "User"
#> 
#> [[6]]$user$user_view_type
#> [1] "public"
#> 
#> [[6]]$user$site_admin
#> [1] FALSE
#> 
#> 
#> [[6]]$labels
#> [[6]]$labels[[1]]
#> [[6]]$labels[[1]]$id
#> [1] -1372889707
#> 
#> [[6]]$labels[[1]]$node_id
#> [1] "LA_kwDOJTnwvM8AAAABritdlQ"
#> 
#> [[6]]$labels[[1]]$url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/labels/deployed"
#> 
#> [[6]]$labels[[1]]$name
#> [1] "deployed"
#> 
#> [[6]]$labels[[1]]$color
#> [1] "9541A4"
#> 
#> [[6]]$labels[[1]]$default
#> [1] FALSE
#> 
#> [[6]]$labels[[1]]$description
#> [1] ""
#> 
#> 
#> 
#> [[6]]$state
#> [1] "open"
#> 
#> [[6]]$locked
#> [1] FALSE
#> 
#> [[6]]$assignee
#> NULL
#> 
#> [[6]]$assignees
#> list()
#> 
#> [[6]]$milestone
#> NULL
#> 
#> [[6]]$comments
#> [1] 4
#> 
#> [[6]]$created_at
#> [1] "2025-09-15T17:23:16Z"
#> 
#> [[6]]$updated_at
#> [1] "2025-09-15T21:21:17Z"
#> 
#> [[6]]$closed_at
#> NULL
#> 
#> [[6]]$author_association
#> [1] "MEMBER"
#> 
#> [[6]]$type
#> NULL
#> 
#> [[6]]$active_lock_reason
#> NULL
#> 
#> [[6]]$sub_issues_summary
#> [[6]]$sub_issues_summary$total
#> [1] 0
#> 
#> [[6]]$sub_issues_summary$completed
#> [1] 0
#> 
#> [[6]]$sub_issues_summary$percent_completed
#> [1] 0
#> 
#> 
#> [[6]]$issue_dependencies_summary
#> [[6]]$issue_dependencies_summary$blocked_by
#> [1] 0
#> 
#> [[6]]$issue_dependencies_summary$total_blocked_by
#> [1] 0
#> 
#> [[6]]$issue_dependencies_summary$blocking
#> [1] 0
#> 
#> [[6]]$issue_dependencies_summary$total_blocking
#> [1] 0
#> 
#> 
#> [[6]]$body
#> [1] "**DO NOT INCLUDE REQUESTS IN THE FIRST COMMENT.**\n**PLEASE POST THIS TEMPLATE UNCHANGED THEN FOLLOW ITS INSTRUCTIONS IN A NEW COMMENT**\n\n\n# General Notes\nThis repository serves as a mostly automated pipeline for deploying workshops to the [Bioconductor Workshop](https://workshop.bioconductor.org).\n\n# Bot Commands\n\nIssues in this repository are listen to by a bot. All commands need to be written on the first line of the comment, in a single line, starting with the `/command` and containing all `parameter=\"value in here\"` pairs.\n\n## Public request\nUnless you are a member of the Bioconductor Core team or an administrator of this repository, you can only perform `/request` commands in this space. Some parameters are universal to all requests and mandatory, namely the Title, Description, Source URL, and Section on the workshop instance. You may request your workshop be added to an existing section, or request a new Section name.\nBelow is an example of an incomplete request with all mandatory parameters:\n```\n/request id=\"myworkshopuniqueid\" title=\"Bold Text\" description=\"unbolded text next to the title\" section=\"Conference 1996\" source=\"https://github.com/super/repository\" \n```\n\nThis request will then need to be completed with an additional 1-2 parameters depending on your source for the workshop.\n\nWe currently support 3 types of submissions:\n\n1) `docker=` parameter for pre-built RStudio containers (such as Orchestra workshops built based on [BuildABiocWorkshop](https://github.com/Bioconductor/BuildABiocWorkshop))\nBelow is an example of a full `/request` command for a workshop with a pre-built docker container:\n```\n/request id=\"tidybioc2022\" title=\"Tidy Transcriptomics\" description=\"For Single-Cell RNA Sequencing Analyses\" section=\"Smorgasbord 2023\" source=\"https://github.com/tidytranscriptomics-workshops/bioc2022_tidytranscriptomics\" docker=\"ghcr.io/tidytranscriptomics-workshops/bioc2022_tidytranscriptomics:latest\"\n```\n\n2) `vignettes=` parameter for Rmd/qmd workshops.\n\nNote that when any `Rmd` or `qmd` files are pulled as part of a request including vignettes, the script will only deploy successfully if the vignettes can successfully render. In the process, all dependencies will be automatically scraped and installed in the resulting container image, making it optional to explicitly mention dependencies.\n\n2) a) In-source vignettes, especially useful for example if your Rmd references other files (eg: images), in which case you should include them in your source repository. Parameter represents a comma-separated list of relative paths (with wildcards accepted) from your source repository. Eg: `vignettes=\"vignettes/*,images/*\"` will copy all files from your source repository under those two subdirectories.\n\nBelow is an example of a full request with in-source vignettes:\n```\n/request id=\"annotation316\" title=\"Genomic Annotation Resources\" description=\"with Bioconductor annotation package\" section=\"Smorgasbord 2023\" source=\"https://github.com/Bioconductor/annotation\" vignettes=\"vignettes/*\"\n```\n\n2) b) markdown file urls, in the form of a comma-separated list of URLs. eg: `vignettes=\"https://raw.githubusercontent.com/Bioconductor/annotation/devel/vignettes/Annotation_Resources.Rmd,https://raw.githubusercontent.com/Bioconductor/annotation/devel/vignettes/Annotating_Genomic_Ranges.Rmd\"`. This can especially be useful for people unfamiliar with github, who could write a self-contained Rmd file in RStudio, paste it at https://gist.github.com where single files can be hosted with no directory structure or `git` operations.\n\nBelow is an example of a full request for a url vignette:\n```\n/request id=\"genranges316\" title=\"Annotating Genomic Ranges\" description=\"with Bioconductor annotation package\" section=\"Smorgasbord 2023\" source=\"https://raw.githubusercontent.com/Bioconductor/annotation/devel/vignettes/Annotating_Genomic_Ranges.Rmd\" vignettes=\"https://raw.githubusercontent.com/Bioconductor/annotation/devel/vignettes/Annotating_Genomic_Ranges.Rmd,https://raw.githubusercontent.com/Bioconductor/annotation/devel/vignettes/Annotating_Genomic_Ranges.Rmd\"\n```\n\n3) `pkglist=` which can be used with or without the `vignettes=` option, in order to add a list of Bioconductor and/or CRAN packages to the resulting auto-built container image for this request. It should be passed as a comma-separated list of package names. eg: `pkglist=\"VariantAnnotation,AnnotationHub,TxDb.Hsapiens.UCSC.hg19.knownGene\"`\n\nBelow is an example of a request to add a package demo\n\n```\n/request id=\"bsgenome316\" title=\"BSgenome package demo\" description=\"in the latest Bioconductor\" section=\"Package Demos\" source=\"https://github.com/Bioconductor/BSgenome\" pkglist=\"BSgenome\"\n```\n\nEvery request must contain at least 1 of the 3 listed parameters, in addition to the mandatory parameters listed above.\n\nYou may add any comments, information or special requests after the second line, but the first line must contain only the `/command` and its parameters.\n\n## Admin request\nBioconductor Core Team and repository admins will be able to deploy vetted requests to one of two servers. These requests take the same parameters as the `/request` command, but go through with building and deploying the changes. When the request does not include a `docker=` parameter, the image building might delay a response by a couple of hours for package-intensive workshops.\n\nThe `/test` command will deploy the instance to our test server, where the requester can verify the aesthetics as well as functionality of the workshop and request any changes.\n\nThe `/publish` command will deploy the final approved workshop to the production instance.\n\nThe best way for an admin to populate these requests is copy-paste the last passing full `/request` command and all parameters from the requester, and replace the `/request` command with the appropriate deployment directive.\n"
#> 
#> [[6]]$closed_by
#> NULL
#> 
#> [[6]]$reactions
#> [[6]]$reactions$url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/112/reactions"
#> 
#> [[6]]$reactions$total_count
#> [1] 0
#> 
#> [[6]]$reactions$`+1`
#> [1] 0
#> 
#> [[6]]$reactions$`-1`
#> [1] 0
#> 
#> [[6]]$reactions$laugh
#> [1] 0
#> 
#> [[6]]$reactions$hooray
#> [1] 0
#> 
#> [[6]]$reactions$confused
#> [1] 0
#> 
#> [[6]]$reactions$heart
#> [1] 0
#> 
#> [[6]]$reactions$rocket
#> [1] 0
#> 
#> [[6]]$reactions$eyes
#> [1] 0
#> 
#> 
#> [[6]]$timeline_url
#> [1] "https://api.github.com/repos/Bioconductor/workshop-contributions/issues/112/timeline"
#> 
#> [[6]]$performed_via_github_app
#> NULL
#> 
#> [[6]]$state_reason
#> NULL
#> 
#>